removed category and display seperated
